refactor - apps now inherit, are not plugins
the original idea was to allow different configurations
to load different plugins, and different apps.

however, I did reach the conclusion that it is easier
to generate a small package that loads the stuff needed
and it makes the code - specially dependencies - easier
to understand.

Additionally, apps often depended on a bunch of plugins,
so there is no point on loading them from a config.

themage